extract install styles into single file
