--- app/components/Approvals/InstallSnapApproval/components/InstallSnapConnectionRequest/InstallSnapConnectionRequest.styles.ts
@@ -1,37 +1,0 @@
-///: BEGIN:ONLY_INCLUDE_IF(snaps)
-import { StyleSheet } from 'react-native';
-import Device from '../../../../../util/device';
-import { Theme } from '../../../../../util/theme/models';
-
-const styleSheet = (params: { theme: Theme }) => {
-  const { theme } = params;
-  const { colors } = theme;
-  return StyleSheet.create({
-    root: {
-      backgroundColor: colors.background.default,
-      paddingTop: 24,
-      borderTopLeftRadius: 20,
-      borderTopRightRadius: 20,
-      minHeight: 200,
-      paddingBottom: Device.isIphoneX() ? 20 : 0,
-    },
-    accountCardWrapper: {
-      paddingHorizontal: 24,
-    },
-    actionContainer: {
-      flex: 0,
-      paddingVertical: 16,
-      justifyContent: 'center',
-    },
-    snapCell: {
-      marginVertical: 16,
-    },
-    description: {
-      textAlign: 'center',
-      paddingBottom: 16,
-    },
-  });
-};
-
-export default styleSheet;
-///: END:ONLY_INCLUDE_IF
